add seperation of concerns before removing code
